requirements.txt
python>=3.7
numpy
torch>=1.14.0
torchvision
matplotlib
tensorflow
tensorboard
pillow